refactor(app): move hilt modules to top level classes
