+++ app/src/free/java/com/github/ashutoshgngwr/noice/di/AnalyticsProviderModule.kt
@@ -1,0 +1,18 @@
+package com.github.ashutoshgngwr.noice.di
+
+import com.github.ashutoshgngwr.noice.provider.AnalyticsProvider
+import com.github.ashutoshgngwr.noice.provider.DummyAnalyticsProvider
+import dagger.Module
+import dagger.Provides
+import dagger.hilt.InstallIn
+import dagger.hilt.components.SingletonComponent
+import javax.inject.Singleton
+
+@Module
+@InstallIn(SingletonComponent::class)
+object AnalyticsProviderModule {
+
+  @Provides
+  @Singleton
+  fun analyticsProvider(): AnalyticsProvider = DummyAnalyticsProvider
+}
