+++ app/src/free/java/com/github/ashutoshgngwr/noice/di/SubscriptionBillingProviderModule.kt
@@ -1,0 +1,21 @@
+package com.github.ashutoshgngwr.noice.di
+
+import com.github.ashutoshgngwr.noice.provider.StripeSubscriptionBillingProvider
+import com.github.ashutoshgngwr.noice.provider.SubscriptionBillingProvider
+import com.trynoice.api.client.NoiceApiClient
+import dagger.Module
+import dagger.Provides
+import dagger.hilt.InstallIn
+import dagger.hilt.components.SingletonComponent
+import javax.inject.Singleton
+
+@Module
+@InstallIn(SingletonComponent::class)
+object SubscriptionBillingProviderModule {
+
+  @Provides
+  @Singleton
+  fun subscriptionProvider(apiClient: NoiceApiClient): SubscriptionBillingProvider {
+    return StripeSubscriptionBillingProvider(apiClient)
+  }
+}
