+++ app/src/main/java/com/github/ashutoshgngwr/noice/NoiceApplication.kt
@@ -1,45 +1,13 @@
 package com.github.ashutoshgngwr.noice
 
 import android.app.Application
-import android.app.PendingIntent
-import android.content.Context
-import android.content.Intent
-import android.os.Build
-import androidx.core.content.ContextCompat
-import androidx.core.content.getSystemService
 import androidx.hilt.work.HiltWorkerFactory
-import androidx.room.Room
 import androidx.work.Configuration
-import com.github.ashutoshgngwr.noice.activity.AlarmRingerActivity
-import com.github.ashutoshgngwr.noice.activity.MainActivity
-import com.github.ashutoshgngwr.noice.data.AppDatabase
-import com.github.ashutoshgngwr.noice.models.Alarm
-import com.github.ashutoshgngwr.noice.repository.AlarmRepository
-import com.github.ashutoshgngwr.noice.repository.PresetRepository
 import com.github.ashutoshgngwr.noice.repository.SettingsRepository
-import com.github.ashutoshgngwr.noice.service.AlarmRingerService
-import com.google.android.exoplayer2.database.DatabaseProvider
-import com.google.android.exoplayer2.database.StandaloneDatabaseProvider
-import com.google.android.exoplayer2.offline.DefaultDownloadIndex
-import com.google.android.exoplayer2.offline.DownloadIndex
-import com.google.android.exoplayer2.offline.WritableDownloadIndex
-import com.google.android.exoplayer2.upstream.cache.Cache
-import com.google.android.exoplayer2.upstream.cache.NoOpCacheEvictor
-import com.google.android.exoplayer2.upstream.cache.SimpleCache
 import com.google.android.material.color.DynamicColors
 import com.google.android.material.color.DynamicColorsOptions
-import com.google.gson.Gson
-import com.google.gson.GsonBuilder
-import com.trynoice.api.client.NoiceApiClient
-import dagger.Module
-import dagger.Provides
-import dagger.hilt.InstallIn
 import dagger.hilt.android.HiltAndroidApp
-import dagger.hilt.android.qualifiers.ApplicationContext
-import dagger.hilt.components.SingletonComponent
-import java.io.File
 import javax.inject.Inject
-import javax.inject.Singleton
 
 @HiltAndroidApp
 class NoiceApplication : Application(), Configuration.Provider {
@@ -62,157 +30,5 @@ class NoiceApplication : Application(), Configuration.Provider {
     return Configuration.Builder()
       .setWorkerFactory(workerFactory)
       .build()
-  }
-
-  @Module
-  @InstallIn(SingletonComponent::class)
-  object GsonModule {
-    @Provides
-    @Singleton
-    fun gson(): Gson = GsonBuilder().create()
-  }
-
-  @Module
-  @InstallIn(SingletonComponent::class)
-  object ApiModule {
-    @Provides
-    @Singleton
-    fun client(@ApplicationContext context: Context, gson: Gson): NoiceApiClient {
-      return NoiceApiClient(
-        context = context,
-        gson = gson,
-        userAgent = "${context.getString(R.string.app_name)}/${BuildConfig.VERSION_NAME} " +
-          "(Android ${Build.VERSION.RELEASE}; ${Build.MANUFACTURER} ${Build.MODEL})",
-      )
-    }
-  }
-
-  @Module
-  @InstallIn(SingletonComponent::class)
-  object AppDatabaseModule {
-    @Provides
-    @Singleton
-    fun appDatabase(@ApplicationContext context: Context): AppDatabase {
-      // TODO: remove in future versions
-      // clear old cache data store.
-      context.deleteDatabase("app-cache.may.db")
-      return Room.databaseBuilder(context, AppDatabase::class.java, "${context.packageName}.db")
-        .build()
-    }
-  }
-
-  @Module
-  @InstallIn(SingletonComponent::class)
-  object ExoPlayerOfflineModule {
-    @Provides
-    @Singleton
-    fun databaseProvider(@ApplicationContext context: Context): DatabaseProvider {
-      return StandaloneDatabaseProvider(context)
-    }
-
-    @Provides
-    @Singleton
-    fun cache(@ApplicationContext context: Context, databaseProvider: DatabaseProvider): Cache {
-      return SimpleCache(
-        File(context.filesDir, "offline-sounds"),
-        NoOpCacheEvictor(),
-        databaseProvider
-      )
-    }
-
-    @Provides
-    @Singleton
-    fun writableDownloadIndex(databaseProvider: DatabaseProvider): WritableDownloadIndex {
-      return DefaultDownloadIndex(databaseProvider)
-    }
-
-    @Provides
-    @Singleton
-    fun readableDownloadIndex(downloadIndex: WritableDownloadIndex): DownloadIndex {
-      return downloadIndex
-    }
-  }
-
-  @Module
-  @InstallIn(SingletonComponent::class)
-  object AlarmRepositoryModule {
-    @Provides
-    @Singleton
-    fun alarmRepository(
-      @ApplicationContext context: Context,
-      appDb: AppDatabase,
-      presetRepository: PresetRepository,
-      settingsRepository: SettingsRepository,
-    ): AlarmRepository {
-      val piFlags = if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.M) {
-        PendingIntent.FLAG_UPDATE_CURRENT or PendingIntent.FLAG_IMMUTABLE
-      } else PendingIntent.FLAG_UPDATE_CURRENT
-
-      val piBuilder = object : AlarmRepository.PendingIntentBuilder {
-        override fun buildShowIntent(alarm: Alarm): PendingIntent {
-          // TODO: focus the specified alarm in the list
-          return Intent(context, MainActivity::class.java)
-            .putExtra(MainActivity.EXTRA_NAV_DESTINATION, R.id.home_alarms)
-            .let { PendingIntent.getActivity(context, alarm.id, it, piFlags) }
-        }
-
-        override fun buildTriggerIntent(alarm: Alarm): PendingIntent {
-          return AlarmRingerService.buildRingIntent(context, alarm.id)
-            .let { intent ->
-              if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.O) {
-                PendingIntent.getForegroundService(context, alarm.id, intent, piFlags)
-              } else {
-                PendingIntent.getService(context, alarm.id, intent, piFlags)
-              }
-            }
-        }
-      }
-
-      return AlarmRepository(
-        requireNotNull(context.getSystemService()),
-        presetRepository,
-        settingsRepository,
-        appDb,
-        piBuilder
-      )
-    }
-  }
-
-  @Module
-  @InstallIn(SingletonComponent::class)
-  object AlarmRingerModule {
-    @Provides
-    @Singleton
-    fun uiController(@ApplicationContext context: Context): AlarmRingerService.UiController {
-      return object : AlarmRingerService.UiController {
-        override fun buildShowIntent(
-          alarmId: Int,
-          alarmLabel: String?,
-          alarmTriggerTime: String,
-        ): Intent {
-          return AlarmRingerActivity.buildIntent(context, alarmId, alarmLabel, alarmTriggerTime)
-        }
-
-        override fun dismiss() {
-          AlarmRingerActivity.dismiss(context)
-        }
-      }
-    }
-
-    @Provides
-    @Singleton
-    fun serviceController(@ApplicationContext context: Context): AlarmRingerActivity.ServiceController {
-      return object : AlarmRingerActivity.ServiceController {
-        override fun dismiss(alarmId: Int) {
-          AlarmRingerService.buildDismissIntent(context, alarmId)
-            .also { ContextCompat.startForegroundService(context, it) }
-        }
-
-        override fun snooze(alarmId: Int) {
-          AlarmRingerService.buildSnoozeIntent(context, alarmId)
-            .also { ContextCompat.startForegroundService(context, it) }
-        }
-      }
-    }
   }
 }
