+++ app/src/test/java/com/github/ashutoshgngwr/noice/receiver/ScheduleExactAlarmPermissionStateChangeReceiverTest.kt
@@ -3,7 +3,7 @@ package com.github.ashutoshgngwr.noice.receiver
 import android.app.AlarmManager
 import android.content.Intent
 import androidx.test.core.app.ApplicationProvider
-import com.github.ashutoshgngwr.noice.NoiceApplication
+import com.github.ashutoshgngwr.noice.di.AlarmRepositoryModule
 import com.github.ashutoshgngwr.noice.repository.AlarmRepository
 import dagger.hilt.android.testing.BindValue
 import dagger.hilt.android.testing.HiltAndroidRule
@@ -20,7 +20,7 @@ import org.junit.runner.RunWith
 import org.robolectric.RobolectricTestRunner
 
 @HiltAndroidTest
-@UninstallModules(NoiceApplication.AlarmRepositoryModule::class)
+@UninstallModules(AlarmRepositoryModule::class)
 @RunWith(RobolectricTestRunner::class)
 class ScheduleExactAlarmPermissionStateChangeReceiverTest {
